% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopxplorer.R
\name{aopOutput}
\alias{aopOutput}
\alias{aopxplorer-shiny}
\alias{renderAop}
\title{Shiny bindings for aopxplorer widgets}
\usage{
aopOutput(outputId, width = "100\%", height = "500px")

renderAop(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates a aopxplorer graph}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using networkD3 widgets within Shiny
applications and interactive Rmd documents.
}

